Move nested profession files into parent directory
We initially had an `add-profession` directory. We'd like to be able to
reuse these files for editing a profession, and there's not a lot of
other admin-style functionality we'd be nesting in similar directories,
so it feels sensible for them to live in the `professions/admin/`
directory.
