+++ src/professions/admin/check-your-answers.controller.spec.ts
@@ -2,15 +2,15 @@ import { DeepMocked, createMock } from '@golevelup/ts-jest';
 import { TestingModule, Test } from '@nestjs/testing';
 import { Request } from 'express';
 import { I18nService } from 'nestjs-i18n';
-import QualificationPresenter from '../../../qualifications/presenters/qualification.presenter';
-import { createMockRequest } from '../../../testutils/create-mock-request';
-import industryFactory from '../../../testutils/factories/industry';
-import legislationFactory from '../../../testutils/factories/legislation';
-import organisationFactory from '../../../testutils/factories/organisation';
-import professionFactory from '../../../testutils/factories/profession';
-import qualificationFactory from '../../../testutils/factories/qualification';
-import { MandatoryRegistration } from '../../profession.entity';
-import { ProfessionsService } from '../../professions.service';
+import QualificationPresenter from '../../qualifications/presenters/qualification.presenter';
+import { createMockRequest } from '../../testutils/create-mock-request';
+import industryFactory from '../../testutils/factories/industry';
+import legislationFactory from '../../testutils/factories/legislation';
+import organisationFactory from '../../testutils/factories/organisation';
+import professionFactory from '../../testutils/factories/profession';
+import qualificationFactory from '../../testutils/factories/qualification';
+import { MandatoryRegistration } from '../profession.entity';
+import { ProfessionsService } from '../professions.service';
 import { CheckYourAnswersController } from './check-your-answers.controller';
 
 describe('CheckYourAnswersController', () => {
