+++ src/professions/admin/confirmation.controller.ts
@@ -8,11 +8,11 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { Response } from 'express';
-import { AuthenticationGuard } from '../../../common/authentication.guard';
-import { ProfessionsService } from '../../professions.service';
+import { AuthenticationGuard } from '../../common/authentication.guard';
+import { ProfessionsService } from '../professions.service';
 import { ConfirmationTemplate } from './interfaces/confirmation.template';
-import { Permissions } from '../../../common/permissions.decorator';
-import { UserPermission } from '../../../users/user.entity';
+import { Permissions } from '../../common/permissions.decorator';
+import { UserPermission } from '../../users/user.entity';
 @UseGuards(AuthenticationGuard)
 @Controller('admin/professions')
 export class ConfirmationController {
