+++ src/professions/admin/dto/regulatory-body.dto.ts
@@ -1,5 +1,5 @@
 import { IsNotEmpty } from 'class-validator';
-import { MandatoryRegistration } from '../../../profession.entity';
+import { MandatoryRegistration } from '../../profession.entity';
 
 export class RegulatoryBodyDto {
   @IsNotEmpty()
