+++ src/professions/admin/interfaces/regulatory-body.template.ts
@@ -1,5 +1,5 @@
-import { RadioButtonArgs } from '../../../../common/interfaces/radio-button-args.interface';
-import { SelectItemArgs } from '../../../../common/interfaces/select-item-args.interface';
+import { RadioButtonArgs } from '../../../common/interfaces/radio-button-args.interface';
+import { SelectItemArgs } from '../../../common/interfaces/select-item-args.interface';
 
 export interface RegulatoryBodyTemplate {
   regulatedAuthoritiesSelectArgs: SelectItemArgs[];
