+++ src/professions/admin/interfaces/top-level-details.template.ts
@@ -1,4 +1,4 @@
-import { CheckboxArgs } from '../../../../common/interfaces/checkbox-args.interface';
+import { CheckboxArgs } from '../../../common/interfaces/checkbox-args.interface';
 
 export interface TopLevelDetailsTemplate {
   name: string | null;
