+++ src/professions/admin/legislation.controller.ts
@@ -10,14 +10,14 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { Response } from 'express';
-import { AuthenticationGuard } from '../../../common/authentication.guard';
-import { Permissions } from '../../../common/permissions.decorator';
-import { Validator } from '../../../helpers/validator';
-import { Legislation } from '../../../legislations/legislation.entity';
-import { UserPermission } from '../../../users/user.entity';
-import { ValidationExceptionFilter } from '../../../common/validation/validation-exception.filter';
-import { ValidationFailedError } from '../../../common/validation/validation-failed.error';
-import { ProfessionsService } from '../../professions.service';
+import { AuthenticationGuard } from '../../common/authentication.guard';
+import { Permissions } from '../../common/permissions.decorator';
+import { Validator } from '../../helpers/validator';
+import { Legislation } from '../../legislations/legislation.entity';
+import { UserPermission } from '../../users/user.entity';
+import { ValidationExceptionFilter } from '../../common/validation/validation-exception.filter';
+import { ValidationFailedError } from '../../common/validation/validation-failed.error';
+import { ProfessionsService } from '../professions.service';
 import LegislationDto from './dto/legislation.dto';
 import { LegislationTemplate } from './interfaces/legislation.template';
