+++ src/professions/professions.module.ts
@@ -3,19 +3,19 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Profession } from './profession.entity';
 import { ProfessionsService } from './professions.service';
 import { ProfessionsController } from './professions.controller';
-import { TopLevelInformationController } from './admin/add-profession/top-level-information.controller';
+import { TopLevelInformationController } from './admin/top-level-information.controller';
 import { IndustriesService } from '../industries/industries.service';
 import { Industry } from '../industries/industry.entity';
-import { CheckYourAnswersController } from './admin/add-profession/check-your-answers.controller';
-import { ConfirmationController } from './admin/add-profession/confirmation.controller';
+import { CheckYourAnswersController } from './admin/check-your-answers.controller';
+import { ConfirmationController } from './admin/confirmation.controller';
 import { SearchController } from './search/search.controller';
-import { RegulatoryBodyController } from './admin/add-profession/regulatory-body.controller';
+import { RegulatoryBodyController } from './admin/regulatory-body.controller';
 import { Organisation } from '../organisations/organisation.entity';
 import { OrganisationsService } from '../organisations/organisations.service';
-import { RegulatedActivitiesController } from './admin/add-profession/regulated-activities.controller';
+import { RegulatedActivitiesController } from './admin/regulated-activities.controller';
 import { ProfessionsController as AdminProfessionsController } from './admin/professions.controller';
-import { QualificationInformationController } from './admin/add-profession/qualification-information.controller';
-import { LegislationController } from './admin/add-profession/legislation.controller';
+import { QualificationInformationController } from './admin/qualification-information.controller';
+import { LegislationController } from './admin/legislation.controller';
 
 @Module({
   imports: [
